Additional reorganization as proposed... I think.
