+++ javascript/opendcs-web-ui-react/src/components/index.tsx
@@ -1,0 +1,2 @@
+export {ColorModes} from "./ColorMode";
+export {ModeIcon, ModeIcons} from "./ModeIcon";
