+++ javascript/opendcs-web-ui-react/src/components/layout/TopBar.tsx
@@ -4,7 +4,7 @@ import { PersonGear } from "react-bootstrap-icons";
 import Nav from "react-bootstrap/Nav";
 import Navbar from "react-bootstrap/Navbar";
 import { AuthContext } from "../../contexts/AuthContext";
-import ColorModes from "../ColorMode";
+import { ColorModes }  from "../";
 
 export function TopBar() {
   const { user, logout } = useContext(AuthContext);
