+++ javascript/opendcs-web-ui-react/src/components/layout/index.tsx
@@ -1,0 +1,2 @@
+export {SideBar} from "./SideBar";
+export {TopBar} from "./TopBar";
