+++ javascript/opendcs-web-ui-react/src/pages/platforms/index.tsx
